[fontloader] patch data-con to use the writable path in potentially writing operation
containers.read() calls caches.loaddata() which at least in the generic
loader (basics-gen) may have the side-effect of byte-compiling a cached
font and saving the result to disk. Since the function is passed the
“readable” paths table it will attempt to write to these paths in order
until a writable one is encountered. Since with a global tex live setup
the first one is not usually writable, this happens on every run. The
log says:

    …
    (compiling luc: /opt/texlive/2016/texmf-var/luatex-cache/generic/fonts/otl/lmroman10-regular.luc)
    (load luc: /home/phg/.texlive2016/texmf-var/luatex-cache/generic/fonts/otl/lmroman10-regular.luc)
    …

Passing the “writable” path instead results in the loader always picking
a usable path first.